apps/chroma_db/requirements.txt
fastapi
uvicorn
chromadb  
pydantic==2.6.4
PyYAML==6.0.1
watchfiles==0.21.0      
psycopg2-binary
sqlalchemy
docx2txt
python-multipart
pymupdf

['aca82611-14fa-4427-bb1b-a2646b5885b9'] ['string'] [{'list_string': ['string']}]